+++ packages/zambdas/ottehr-spec.json
@@ -392,8 +392,8 @@
       "src": "src/ehr/get-create-lab-order-resources/index",
       "zip": ".dist/zips/get-create-lab-order-resources.zip"
     },
-    "CREATE_REVIEW_LAB_RESULTS_TASK": {
-      "name": "CREATE_REVIEW_LAB_RESULTS_TASK",
+    "CREATE-REVIEW-LAB-RESULTS-TASK": {
+      "name": "CREATE-REVIEW-LAB-RESULT-TASK",
       "type": "subscription",
       "runtime": "nodejs20.x",
       "src": "src/subscriptions/diagnostic-report/create-review-lab-results-task/index",
@@ -832,13 +832,6 @@
       "runtime": "nodejs20.x",
       "src": "src/ehr/submit-lab-order/index",
       "zip": ".dist/zips/submit-lab-order.zip"
-    },
-    "RESULTS-LAB-ORDER": {
-      "name": "RESULTS-LAB-ORDER",
-      "type": "http_auth",
-      "runtime": "nodejs20.x",
-      "src": "src/ehr/results-lab-order/index",
-      "zip": ".dist/zips/results-lab-order.zip"
     },
     "UPDATE-LAB-ORDER-RESOURCES": {
       "name": "UPDATE-LAB-ORDER-RESOURCES",
